requirements.txt
dill==0.3.1.1
Flask==1.1.2
numpy==1.16.4
Pillow==7.1.2
scikit_learn==0.22
scikit-image==0.16.2
tensorflow==2.0.0
setuptools<45.0.0
pyinstaller==3.5
